# multimodule/multi-swagger/doc/html/swagger.json
{"swagger":"2.0","info":{"description":"简单优雅的restfun风格，http://blog.csdn.net/forezp","version":"1.0","title":"springboot利用swagger构建api文档","contact":{"name":"Hao Lin","url":"http://wiki.virtueit.net/display/CMDT/CARDAY+Backend+vService+API+Spec","email":"andy_linhao@outlook.com"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"BookApi","description":"图书基本信息操作API"},{"name":"StudentApi","description":"学生基本信息操作API"}],"paths":{"/books":{"get":{"tags":["BookApi"],"summary":"获取图书列表","description":"获取图书列表","operationId":"getBookUsingGET_1","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Book"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["BookApi"],"summary":"创建图书","description":"创建图书","operationId":"postBookUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"book","description":"图书详细实体","required":true,"schema":{"$ref":"#/definitions/Book"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/books/{id}":{"get":{"tags":["BookApi"],"summary":"获图书细信息","description":"根据url的id来获取详细信息","operationId":"getBookUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"ID","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Book"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["BookApi"],"summary":"更新信息","description":"根据url的id来指定更新图书信息","operationId":"putUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"图书ID","required":true,"type":"integer","format":"int64"},{"in":"body","name":"book","description":"图书实体book","required":true,"schema":{"$ref":"#/definitions/Book"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["BookApi"],"summary":"删除图书","description":"根据url的id来指定删除图书","operationId":"deleteUserUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"图书ID","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/student":{"get":{"tags":["StudentApi"],"summary":"getStudent","description":"依据学生姓名查询学生信息","operationId":"getStudentUsingGET","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Student"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["StudentApi"],"summary":"addStudent","description":"添加一个学生","operationId":"addStudentUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"student","description":"student","required":true,"schema":{"$ref":"#/definitions/Student"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Book":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number","format":"double"}}},"Student":{"type":"object","properties":{"address":{"type":"string","description":"家庭住址"},"age":{"type":"integer","format":"int32","description":"年龄"},"cls":{"type":"string","description":"班级"},"id":{"type":"integer","format":"int32","description":"学号"},"name":{"type":"string","description":"姓名"},"sex":{"type":"string","description":"性别"}},"description":"学生信息描述"}}}